Första knappen i gränssnittet fungerar

+++ Connection/DataAccessLayer.cs
@@ -45,7 +45,7 @@ namespace Connection
             reader.Close();
 
         }
-        public void insertProduct(int productId, string productName, double productPrice, int categoryId)
+        public void insertProduct(int productId, string productName, decimal productPrice, int categoryId)
         {
             SqlConnection connection = GetDatabaseConnection();
             SqlCommand command = connection.CreateCommand();
